+++ react-components/src/views/ProductsModal/StepCategory.jsx
@@ -4,15 +4,9 @@ import PropTypes from 'prop-types'
 
 import Services from '@src/Services'
 import Field from '@src/components/Field'
-import SocialLoginButtons from '@src/components/SocialLoginButtons'
 
 
 export default function StepCategory(props){
-
-  function handleChange(value) {
-    props.handleChange(value)
-    props.handleSubmit()
-  }
 
   return (
     <div className='great-mvp-export-wizard-step-category'>
@@ -22,7 +16,7 @@ export default function StepCategory(props){
         name="category"
         disabled={props.disabled}
         value={props.value}
-        handleChange={handleChange}
+        handleChange={props.handleChange}
         autofocus={true}
         errors={props.errors.category || []}
         options={[
@@ -39,7 +33,6 @@ StepCategory.propTypes = {
   disabled: PropTypes.bool,
   errors: PropTypes.object,
   handleChange: PropTypes.func.isRequired,
-  handleSubmit: PropTypes.func.isRequired,
   value: PropTypes.string,
 }